refactor(api,models,client,session,http,error)!: restructure API, add typed models, new session/http, and rename crate
- Rename crate to `degiro-ox` in `Cargo.toml`.
- Replace ad-hoc API responses with strongly typed models under `src/models/*` (account, portfolio, product, money, risk, periods, financial reports, etc.).
- Introduce `src/http.rs` for centralized HTTP client, cookie handling, and request utilities; add `backon` for retry/backoff and `tracing` hooks.
- Add `src/session.rs` for auth/session lifecycle, 2FA/TOTP support, and credential handling; adjust `src/client.rs` to use new session/http abstractions.
- Consolidate and simplify API endpoints in `src/api/*`, removing duplication, aligning signatures, and leveraging shared serde helpers (`src/serde_utils.rs`).
- Overhaul error handling in `src/error.rs` with `thiserror`, richer variants, and better propagation across API layers.
- Introduce decimal and crypto libs for money and secure storage: `rust_decimal`, `aes-gcm`, `argon2`, `rand`.
- Remove deprecated utilities (`src/util.rs`, old `src/money.rs`) in favor of new typed modules.
- Add planning/docs and reviews: `refactor_plan.md`, `todo/*`, `code_review/*`, and `unexpected_error_analysis.md`.

BREAKING CHANGE:
- Public API changed across `api/*`, `client`, and models; downstream callers must update to new types and function signatures.
- Crate renamed to `degiro-ox`.

Affects: src/api/*, src/models/*, src/{client,session,http,error,lib}.rs, Cargo.toml

+++ src/models/order.rs
@@ -3,7 +3,6 @@ use std::{collections::HashSet, fmt};
 use chrono::{DateTime, Utc};
 use derivative::Derivative;
 use derive_more::derive::{Deref, DerefMut};
-use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use strum::{Display, EnumString};
 
@@ -19,14 +18,14 @@ pub struct Order {
     pub product_id: u64,
     pub product: String,
     pub contract_type: u64,
-    pub contract_size: Decimal,
+    pub contract_size: f64,
     pub currency: Currency,
     pub transaction_type: TransactionType,
-    pub size: Decimal,
-    pub quantity: Decimal,
-    pub price: Decimal,
-    pub stop_price: Decimal,
-    pub total_order_value: Decimal,
+    pub size: f64,
+    pub quantity: f64,
+    pub price: f64,
+    pub stop_price: f64,
+    pub total_order_value: f64,
     pub order_type: OrderType,
     pub order_type_id: u8,
     pub order_time_type: OrderTimeType,
@@ -89,7 +88,7 @@ impl Orders {
         Orders(orders)
     }
 
-    pub fn total_value(&self) -> Decimal {
+    pub fn total_value(&self) -> f64 {
         self.iter().map(|o| o.total_order_value).sum()
     }
